refactor: Completely modularize hook library
