+++ src-hook/src/hooks/sba.rs
@@ -1,0 +1,241 @@
+use std::sync::atomic::Ordering;
+
+use anyhow::{anyhow, Result};
+use protocol::Message;
+use retour::static_detour;
+
+use crate::{event, process::Process};
+
+use super::globals::SBA_OFFSET;
+
+type OnSBAUpdateFunc = unsafe extern "system" fn(*const usize, f32, u32, u8, u32, u8) -> usize;
+type OnSBAAttemptFunc = unsafe extern "system" fn(*const usize, f32) -> usize;
+type OnCheckSBACollisionFunc = unsafe extern "system" fn(*const usize, f32) -> usize;
+type OnContinueSBAChainFunc = unsafe extern "system" fn(*const usize, *const usize) -> usize;
+
+static_detour! {
+    static OnSBAUpdate: unsafe extern "system" fn(*const usize, f32, u32, u8, u32, u8) -> usize;
+    static OnSBAAttempt: unsafe extern "system" fn(*const usize, f32) -> usize;
+    static OnCheckSBACollision: unsafe extern "system" fn(*const usize, f32) -> usize;
+    static OnContinueSBAChain: unsafe extern "system" fn(*const usize, *const usize) -> usize;
+}
+
+const ON_HANDLE_SBA_UPDATE_SIG: &str = "e8 $ { ' } c5 fa 10 46 ? c5 f8 2e 86 80 00 00 00";
+const ON_ATTEMPT_SBA_SIG: &str = "e8 $ { ' } 48 8d 8e a0 9c ff ff c7 44 24 38 00 00 80 3f";
+const ON_CHECK_SBA_COLLISION_SIG: &str = "e8 $ { ' } 84 c0 0f 85 f0 00 00 ? 8b 8e a8 9c ff ff";
+const ON_CONTINUE_SBA_CHAIN_SIG: &str = "e8 $ { ' } 48 8b 53 ? 48 8d 82 ? ? ? ?";
+
+/// Gets called when your SBA gauge value needs to update with a given value.
+#[derive(Clone)]
+pub struct OnHandleSBAUpdateHook {
+    tx: event::Tx,
+}
+
+impl OnHandleSBAUpdateHook {
+    pub fn new(tx: event::Tx) -> Self {
+        OnHandleSBAUpdateHook { tx }
+    }
+
+    pub fn setup(&self, process: &Process) -> Result<()> {
+        if let Ok(on_sba_update_original) = process.search_address(ON_HANDLE_SBA_UPDATE_SIG) {
+            #[cfg(feature = "console")]
+            println!("found on sba update");
+
+            let cloned_self = self.clone();
+
+            unsafe {
+                let func: OnSBAUpdateFunc = std::mem::transmute(on_sba_update_original);
+                OnSBAUpdate.initialize(func, move |a1, a2, a3, a4, a5, a6| {
+                    cloned_self.run(a1, a2, a3, a4, a5, a6)
+                })?;
+                OnSBAUpdate.enable()?;
+            }
+        } else {
+            return Err(anyhow!("Could not find on_sba_update"));
+        }
+
+        Ok(())
+    }
+
+    fn run(&self, a1: *const usize, a2: f32, a3: u32, a4: u8, a5: u32, a6: u8) -> usize {
+        let sba_offset = SBA_OFFSET.load(Ordering::Relaxed);
+
+        let entity_ptr = unsafe { a1.byte_sub(sba_offset as usize) } as *const usize;
+        let sba_value_ptr = unsafe { a1.byte_add(0x7C) } as *const f32;
+
+        let old_sba_value = unsafe { sba_value_ptr.read() };
+
+        let ret = unsafe { OnSBAUpdate.call(a1, a2, a3, a4, a5, a6) };
+
+        let player_idx = unsafe { entity_ptr.byte_add(0x170).read() } as u32;
+        let new_sba_value = unsafe { sba_value_ptr.read() };
+        let sba_added = f32::max(new_sba_value - old_sba_value, 0.0);
+
+        let payload = Message::OnUpdateSBA(protocol::OnUpdateSBAEvent {
+            actor_index: player_idx,
+            sba_value: new_sba_value,
+            sba_added,
+        });
+
+        let _ = self.tx.send(payload);
+
+        ret
+    }
+}
+
+/// Called when your first try to attempt your SBA, and sets you into "casting SBA" state.
+#[derive(Clone)]
+pub struct OnAttemptSBAHook {
+    tx: event::Tx,
+}
+
+impl OnAttemptSBAHook {
+    pub fn new(tx: event::Tx) -> Self {
+        OnAttemptSBAHook { tx }
+    }
+
+    pub fn setup(&self, process: &Process) -> Result<()> {
+        if let Ok(on_sba_attempt_original) = process.search_address(ON_ATTEMPT_SBA_SIG) {
+            #[cfg(feature = "console")]
+            println!("found on sba attempt");
+
+            let cloned_self = self.clone();
+
+            unsafe {
+                let func: OnSBAAttemptFunc = std::mem::transmute(on_sba_attempt_original);
+                OnSBAAttempt.initialize(func, move |a1, a2| cloned_self.run(a1, a2))?;
+                OnSBAAttempt.enable()?;
+            }
+        } else {
+            return Err(anyhow!("Could not find on_sba_attempt"));
+        }
+
+        Ok(())
+    }
+
+    fn run(&self, a1: *const usize, a2: f32) -> usize {
+        let ret = unsafe { OnSBAAttempt.call(a1, a2) };
+
+        let entity_ptr = unsafe { a1.byte_add(0x10).read() } as *const usize;
+        let player_index = unsafe { entity_ptr.byte_add(0x170).read() } as u32;
+
+        #[cfg(feature = "console")]
+        println!("on sba attempt: player_index={}", player_index);
+
+        let payload = Message::OnAttemptSBA(protocol::OnAttemptSBAEvent {
+            actor_index: player_index,
+        });
+
+        let _ = self.tx.send(payload);
+
+        ret
+    }
+}
+
+/// Gets called when you're in "casting SBA state" once per game update interval until your SBA lands on
+/// the target (or you miss)
+#[derive(Clone)]
+pub struct OnCheckSBACollisionHook {
+    tx: event::Tx,
+}
+
+impl OnCheckSBACollisionHook {
+    pub fn new(tx: event::Tx) -> Self {
+        OnCheckSBACollisionHook { tx }
+    }
+
+    pub fn setup(&self, process: &Process) -> Result<()> {
+        if let Ok(on_check_sba_collision_original) =
+            process.search_address(ON_CHECK_SBA_COLLISION_SIG)
+        {
+            #[cfg(feature = "console")]
+            println!("found on check sba collision");
+
+            let cloned_self = self.clone();
+
+            unsafe {
+                let func: OnCheckSBACollisionFunc =
+                    std::mem::transmute(on_check_sba_collision_original);
+                OnCheckSBACollision.initialize(func, move |a1, a2| cloned_self.run(a1, a2))?;
+                OnCheckSBACollision.enable()?;
+            }
+        } else {
+            return Err(anyhow!("Could not find on_check_sba_collision"));
+        }
+
+        Ok(())
+    }
+
+    fn run(&self, a1: *const usize, a2: f32) -> usize {
+        let ret = unsafe { OnCheckSBACollision.call(a1, a2) };
+
+        if ret != 0 {
+            let entity_ptr = unsafe { a1.byte_add(0x10).read() } as *const usize;
+            let player_index = unsafe { entity_ptr.byte_add(0x170).read() } as u32;
+
+            #[cfg(feature = "console")]
+            println!("on perform sba: player_index={}", player_index);
+
+            let payload = Message::OnPerformSBA(protocol::OnPerformSBAEvent {
+                actor_index: player_index,
+            });
+
+            let _ = self.tx.send(payload);
+        }
+
+        ret
+    }
+}
+
+/// Gets called when you connect your SBA with an active SBA chain (2/3/4)
+#[derive(Clone)]
+pub struct OnContinueSBAChainHook {
+    tx: event::Tx,
+}
+
+impl OnContinueSBAChainHook {
+    pub fn new(tx: event::Tx) -> Self {
+        OnContinueSBAChainHook { tx }
+    }
+
+    pub fn setup(&self, process: &Process) -> Result<()> {
+        if let Ok(on_continue_sba_chain_original) =
+            process.search_address(ON_CONTINUE_SBA_CHAIN_SIG)
+        {
+            #[cfg(feature = "console")]
+            println!("found on continue sba chain");
+
+            let cloned_self = self.clone();
+
+            unsafe {
+                let func: OnContinueSBAChainFunc =
+                    std::mem::transmute(on_continue_sba_chain_original);
+                OnContinueSBAChain.initialize(func, move |a1, a2| cloned_self.run(a1, a2))?;
+                OnContinueSBAChain.enable()?;
+            }
+        } else {
+            return Err(anyhow!("Could not find on_continue_sba_chain"));
+        }
+
+        Ok(())
+    }
+
+    fn run(&self, player_entity: *const usize, a2: *const usize) -> usize {
+        #[cfg(feature = "console")]
+        println!(
+            "on continue sba chain: player_entity={:p}, a2={:p}",
+            player_entity, a2
+        );
+
+        let ret = unsafe { OnContinueSBAChain.call(player_entity, a2) };
+        let player_index = unsafe { player_entity.byte_add(0x170).read() } as u32;
+
+        let payload = Message::OnContinueSBAChain(protocol::OnContinueSBAChainEvent {
+            actor_index: player_index,
+        });
+
+        let _ = self.tx.send(payload);
+
+        ret
+    }
+}
